refactor: add constants, update 3d mimetype
