+++ docarray/typing/url/video_url.py
@@ -4,6 +4,7 @@ from typing import List, Optional, TypeVar
 from docarray.typing.bytes.video_bytes import VideoBytes, VideoLoadResult
 from docarray.typing.proto_register import _register_proto
 from docarray.typing.url.any_url import AnyUrl
+from docarray.typing.url.mimetypes import VIDEO_MIMETYPE
 from docarray.utils._internal.misc import is_notebook
 
 T = TypeVar('T', bound='VideoUrl')
@@ -18,7 +19,7 @@ class VideoUrl(AnyUrl):
 
     @classmethod
     def mime_type(cls) -> str:
-        return 'video'
+        return VIDEO_MIMETYPE
 
     @classmethod
     def extra_extensions(cls) -> List[str]:
